requirements.txt
pandas>=2.0.0
plotly>=5.0.0
dash>=2.0.0
fredapi>=0.5
python-dotenv==1.0.1